# GSIE/API/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "GoogleLoginRequest": {
        "additionalProperties": false,
        "description": "Preuve Google OIDC liée à un nonce serveur à usage unique.",
        "properties": {
          "id_token": {
            "format": "password",
            "maxLength": 16384,
            "minLength": 1,
            "title": "Id Token",
            "type": "string",
            "writeOnly": true
          },
          "nonce": {
            "format": "password",
            "maxLength": 256,
            "minLength": 32,
            "title": "Nonce",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "id_token",
          "nonce"
        ],
        "title": "GoogleLoginRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LocalLoginRequest": {
        "additionalProperties": false,
        "description": "Connexion par adresse e-mail et mot de passe.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LocalLoginRequest",
        "type": "object"
      },
...
      "RegistrationRequest": {
        "additionalProperties": false,
        "description": "Création d'un compte Quintessences local.",
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 12,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "RegistrationRequest",
        "type": "object"
      },
...
      "TokenResponse": {
        "additionalProperties": false,
        "description": "Réponse token — access + refresh.",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "description": "Durée de vie du access token en secondes",
            "title": "Expires In",
            "type": "integer"
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "token_type": {
            "default": "Bearer",
            "title": "Token Type",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_in"
        ],
        "title": "TokenResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "GSIE — General System Intelligence Engine API",
    "title": "GSIE API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/link/google": {
      "post": {
        "operationId": "link_google_api_v1_auth_link_google_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Rattacher Google au compte courant",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/login/google": {
      "post": {
        "operationId": "login_google_api_v1_auth_login_google_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Se connecter avec Google",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login/password": {
      "post": {
        "description": "Authentifie sans distinguer compte absent et mot de passe erroné.",
        "operationId": "login_local_api_v1_auth_login_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Se connecter par adresse e-mail",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/register": {
      "post": {
        "description": "Crée le compte canonique et son premier moyen de connexion.",
        "operationId": "register_local_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Créer un compte Quintessences local",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
...
}